reorganized the response parsing classes
