--- practice-app/backend/disasterresponse/src/main/java/Response/ImdbResponse/Image.java
@@ -1,20 +1,0 @@
-package Response.ImdbResponse;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Image {
-    private int height;
-    private String id;
-    private String url;
-    private int width;
-
-}
-
-
