--- practice-app/backend/disasterresponse/src/main/java/Response/ImdbResponse/Role.java
@@ -1,16 +1,0 @@
-package Response.ImdbResponse;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Role {
-    private String character;
-    private String characterId;
-
-}
