+++ practice-app/backend/disasterresponse/src/main/java/com/a1/disasterresponse/model/ImdbResponse.java
@@ -1,11 +1,10 @@
-package Response.ImdbResponse;
+package com.a1.disasterresponse.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.NoArgsConstructor;
-
 import java.util.List;
