+++ practice-app/backend/disasterresponse/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 spring.datasource.url = jdbc:postgresql://localhost:5432/disasterresponse
 spring.datasource.driverClassName = org.postgresql.Driver
 spring.datasource.username =postgres
-spring.datasource.password =112212
+spring.datasource.password =furkan
 spring.jpa.hibernate.ddl-auto= create
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
